refactor: rename InputField to FormField and make it configurable with other components
