+++ src/components/FormField/index.tsx
@@ -1,0 +1,31 @@
+import { colors, fontSizes } from "@/src/common/styleVariables";
+import styled from "@emotion/styled";
+import { ComponentProps, ComponentType, ReactNode } from "react";
+import Input from "../Input";
+
+const Label = styled.label({
+	fontSize: fontSizes.small,
+});
+
+const Error = styled.span({
+	color: colors.error,
+});
+
+type Props<Component extends ComponentType> = ComponentProps<Component> & {
+	component?: Component;
+	label: ReactNode;
+	error?: string;
+};
+
+const DEFAULT_COMPONENT = Input;
+
+export default function FormField<Component extends ComponentType = typeof DEFAULT_COMPONENT>(props: Props<Component>) {
+	const { component: Component = DEFAULT_COMPONENT, label, error, ...otherProps } = props;
+	return (
+		<Label>
+			{label}
+			<Component error={error} {...otherProps} />
+			{error && <Error aria-live="assertive">{error}</Error>}
+		</Label>
+	);
+}
